Major_project_Medical_Healthcare/project/database.py
import os
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv

load_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), ".env"))

MONGODB_URI = os.getenv("MONGODB_URI")

client = AsyncIOMotorClient(MONGODB_URI)
db = client["medical_healthcare_db"]

# Collections
users_collection = db["users"]
chat_history_collection = db["chat_history"]
