Major_project_Medical_Healthcare/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
httpx
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pymongo[srv]
requests